+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- when calling `BinRemapper::new`, the limits are now checked for overlaps, in
+  which case a new error is returned
+
 ## [0.6.3] - 12/12/2023
 
 ### Added